lib clean salloc etc

+++ ft_strlcat.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gbeco <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: gbeco <gbeco@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 13:58:40 by gbeco             #+#    #+#             */
-/*   Updated: 2020/11/25 16:26:24 by gbeco            ###   ########lyon.fr   */
+/*   Updated: 2021/06/02 10:55:29 by gbeco            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 	size_t	len;
 
 	len = ft_strlen(src) + ft_strlen(dst);
-	if (ft_strlen(dst) > dstsize - 1 || dstsize == 0)
+	if ((size_t)ft_strlen(dst) > dstsize - 1 || dstsize == 0)
 		return (dstsize + ft_strlen(src));
 	i = 0;
 	while (dst[i])
